+++ core/src/test/scala/com/coxautodata/vegalite4s/PlotHelpersSpec.scala
@@ -1,14 +1,18 @@
 package com.coxautodata.vegalite4s
 
+import java.sql.{Date, Timestamp}
+
 import com.coxautodata.vegalite4s.PlotHelpers._
 import org.scalatest.{FunSpec, Matchers}
+import java.math.{BigDecimal => JBigDecimal, BigInteger => JBigInteger}
 
 class PlotHelpersSpec extends FunSpec with Matchers {
 
   it("add a title to a plot") {
 
     val plot = VegaLite()
-      .withObject("""
+      .withObject(
+        """
           |{
           |  "description": "A simple bar chart with embedded data.",
           |  "mark": "bar",
@@ -43,7 +47,8 @@ class PlotHelpersSpec extends FunSpec with Matchers {
   it("add data to a plot") {
 
     val plot = VegaLite()
-      .withObject("""
+      .withObject(
+        """
           |{
           |  "description": "A simple bar chart with embedded data.",
           |  "mark": "bar",
@@ -103,7 +108,8 @@ class PlotHelpersSpec extends FunSpec with Matchers {
   it("add width to a plot") {
 
     val plot = VegaLite()
-      .withObject("""
+      .withObject(
+        """
           |{
           |  "description": "A simple bar chart with embedded data.",
           |  "mark": "bar",
@@ -138,7 +144,8 @@ class PlotHelpersSpec extends FunSpec with Matchers {
   it("add height to a plot") {
 
     val plot = VegaLite()
-      .withObject("""
+      .withObject(
+        """
           |{
           |  "description": "A simple bar chart with embedded data.",
           |  "mark": "bar",
@@ -170,4 +177,73 @@ class PlotHelpersSpec extends FunSpec with Matchers {
 
   }
 
+  describe("anyEncoder") {
+
+    it("encode Int") {
+      anyEncoder(1).noSpaces should be("1")
+    }
+
+    it("encode Long") {
+      anyEncoder(1L).noSpaces should be("1")
+    }
+
+    it("encode Short") {
+      anyEncoder(1.toShort).noSpaces should be("1")
+    }
+
+    it("encode Boolean") {
+      anyEncoder(true).noSpaces should be("true")
+    }
+
+    it("encode String") {
+      anyEncoder("string").noSpaces should be("\"string\"")
+    }
+
+    it("encode BigInt") {
+      anyEncoder(BigInt("123456789123456789")).noSpaces should be("123456789123456789")
+    }
+
+    it("encode Java BigInt") {
+      anyEncoder(new JBigInteger("123456789123456789")).noSpaces should be("123456789123456789")
+    }
+
+    it("encode double") {
+      anyEncoder(1.3d).noSpaces should be("1.3")
+    }
+
+    it("encode float") {
+      anyEncoder(1.3f).noSpaces should be("1.3")
+    }
+
+    it("encode BigDecimal") {
+      anyEncoder(BigDecimal("123456789123456789.1")).noSpaces should be("123456789123456789.1")
+    }
+
+    it("encode Java BigDecimal") {
+      anyEncoder(new JBigDecimal("123456789123456789.1")).noSpaces should be("123456789123456789.1")
+    }
+
+    it("encode Timestamp") {
+      anyEncoder(new Timestamp(1552306550229L)).noSpaces should be("\"2019-03-11T12:15:50.229\"")
+    }
+
+    it("encode Date") {
+      anyEncoder(new Date(1552306550229L)).noSpaces should be("\"2019-03-11\"")
+    }
+
+    it("encode null") {
+      anyEncoder(null).noSpaces should be("null")
+    }
+
+    it("encode TestObject") {
+      anyEncoder(TestObject).noSpaces should be("\"TestValue\"")
+    }
+
+  }
+
+
+}
+
+object TestObject {
+  override def toString: String = "TestValue"
 }