+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Fixed
+- Null and Java BigInteger and BigDecimal datatype conversions
+
 ## v0.1 - 2019-02-18
 
 ### Added